Add getter methods to ensure OOP encapsulation principle
