+++ animals.rb
@@ -1,0 +1,32 @@
+class Animal
+  def initialize(type, number_of_legs, name = "Unknown")
+    @id = Random.rand(1..1000)
+    @name = name
+    @number_of_legs = number_of_legs
+    @type=type
+  end
+
+  def id
+    @id
+  end
+
+  def type
+    @type
+  end
+
+  def number_of_legs
+    @number_of_legs
+  end
+
+  def name
+    @name
+  end
+
+  def name=(value)
+    @name = value
+  end
+
+  def speak
+    "Bla bla bla"
+  end
+end
