+++ main.rb
@@ -1,18 +1,12 @@
-class Animal
-  def initialize(number_of_legs, name = "Unknown")
-    @id = Random.rand(1..1000)
-    @name = name
-    @number_of_legs = number_of_legs
-  end
+require "./animals.rb"
 
-  def speak
-    "Bla bla bla"
-  end
-end
+p animal_1 = Animal.new("dog", 4, "Rex")
+p animal_1.id
+p animal_1.type
+p animal_1.name
+p animal_1.number_of_legs
 
-animal_1 = Animal.new(4, "Rex")
-animal_2 = Animal.new(8)
-
-
-p animal_1.speak
-p animal_2.speak
+p animal_2 = Animal.new("cat", 8)
+p animal_2.name
+p animal_2.name = "Fluffy"
+p animal_2.name
